+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Bump data-export interface version to 5.0.
 * Revert elastic search. Refs UIIN-1822.
 * Use correct `css-loader` syntax. Refs UIIN-1826.
+* Fix suppressed from discovery filter. Fixes UIIN-1832.
 
 ## [8.0.0](https://github.com/folio-org/ui-inventory/tree/v8.0.0) (2021-10-05)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v7.1.4...v8.0.0)